refactor(backend): resolve circular dependency in user types
