+++ backend/src/modules/auth/schemas/index.ts
@@ -4,7 +4,7 @@ import { z } from 'zod';
 extendZodWithOpenApi(z);
 
 // Password schema
-const _passwordSchema = z
+const passwordSchema = z
     .string()
     .min(8, 'Password must be at least 8 characters')
     .regex(/[A-Z]/, 'Password must include an uppercase letter')
@@ -12,26 +12,10 @@ const _passwordSchema = z
     .regex(/\d+/, 'Password must include a number')
     .regex(/[!@#$%^&*()_+\-=[\]{};':"\\|,.<>/?~`]/, 'Password must include a special character');
 
-// Register schema
-const registerUserPayloadSchema = z
-    .object({
-        firstName: z.string(),
-        lastName: z.string(),
-        email: z.string().email(),
-        password: _passwordSchema,
-        confirmationPassword: _passwordSchema,
-    })
-    // Check if passwords match
-    .refine(data => data.password === data.confirmationPassword, {
-        path: ['confirmationPassword'],
-        message: 'Passwords do not match',
-    })
-    .openapi('RegisterUserPayload');
-
 // Login schema
 const loginUserPayloadSchema = z
     .object({
-        password: z.string(),
+        password: passwordSchema,
         email: z.string().email(),
     })
     .openapi('LoginUserPayload');
@@ -40,11 +24,13 @@ const loginUserPayloadSchema = z
 const accessTokenSchema = z.string().jwt().openapi('AccessToken');
 
 // JWT payload schema
-const jwtPayloadSchema = z.object({
-    firstName: z.string(),
-    lastName: z.string(),
-    email: z.string().email(),
-    id: z.string(),
-});
+const jwtPayloadSchema = z
+    .object({
+        firstName: z.string(),
+        lastName: z.string(),
+        email: z.string().email(),
+        id: z.string(),
+    })
+    .openapi('JwtPayload');
 
-export { accessTokenSchema, jwtPayloadSchema, loginUserPayloadSchema, registerUserPayloadSchema };
+export { accessTokenSchema, jwtPayloadSchema, loginUserPayloadSchema, passwordSchema };
