+++ backend/src/modules/shared/types/user/index.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { registerUserPayloadSchema } from 'modules/auth/schemas';
+import { registerUserPayloadSchema } from '../../schemas/user';
 
 type RegisterUserPayload = z.infer<typeof registerUserPayloadSchema>;
 type CreateUserPayload = Omit<RegisterUserPayload, 'confirmationPassword'>;
